parser_service/requirements.txt
fastapi
uvicorn
python-multipart
unstructured
unstructured[pptx]
unstructured[pdf]
pymupdf
sentence-transformers
pydantic-settings